requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
idna==3.4
oauthlib==3.2.2
python-dotenv==0.21.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tweepy==4.12.1
tzdata==2022.7
urllib3==1.26.14
